requirements.txt
# Python version: 3.8.10
smbus2==0.4.3
bmp280==1.0.0
Adafruit-SSD1306==1.6.2
Pillow==8.1.2
RPi.GPIO==0.7.0